requirements.txt
apache-airflow
playwright
python-dotenv
openai
google-generativeai
pytest
fastapi
uvicorn
httpx
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
